restructured again - oops. Learning so much about packaging :)
