+++ .gitignore
@@ -2,5 +2,5 @@
 .*.swp
 venv
 .pytest_cache
-__pycache__
+.*__pycache__
 dist
