--- ansifier/__init__.py
@@ -1,2 +1,0 @@
-from .image_printer import ImageFilePrinter
-from .config import UTF8_DEFAULT_CHARS, UTF8_BLOCK_CHARS, ASCII_DEFAULT_CHARS
