+++ makefile
@@ -2,4 +2,5 @@ main:
 	python -m build
 
 test:
-	PYTHONPATH=./ansifier python -m pytest
+	python -m pytest
+
